More unboxing cleanup.
git-svn-id: https://svn.code.sf.net/p/vassalengine/svn/VASSAL-src/trunk@9393 67b53d14-2c14-4ace-a08f-0dab2b34000c

+++ src/VASSAL/chat/CgiServerStatus.java
@@ -99,7 +99,7 @@ public class CgiServerStatus implements ServerStatus {
 
   public ModuleSummary[] getHistory(String timeRange) {
     final Long l = timeRanges.get(timeRange);
-    return l != null ? getHistory(l.longValue()) : new ModuleSummary[0];
+    return l != null ? getHistory(l) : new ModuleSummary[0];
   }
 
   public String[] getSupportedTimeRanges() {
